# Conflicts:
#	app/src/main/java/com/gat/app/screen/ParcelableScreen.java
#	app/src/main/java/com/gat/app/screen/ScreenPresenterFactory.java
#	app/src/main/java/com/gat/data/user/UserDataSourceImpl.java
#	app/src/main/java/com/gat/dependency/PresenterComponent.java
#	app/src/main/java/com/gat/dependency/PresenterModule.java
#	app/src/main/java/com/gat/feature/scanbarcode/ScanBarcodeActivity.java
#	app/src/main/java/com/gat/feature/start/StartActivity.java
#	app/src/main/java/com/gat/repository/datasource/UserDataSource.java

+++ app/src/main/java/com/gat/feature/personal/adapter/BookSharingAdapter.java
@@ -71,7 +71,7 @@ public class BookSharingAdapter extends  RecyclerView.Adapter<BookSharingAdapter
             if(!Strings.isNullOrEmpty(entity.getBorrowingUserName())){
                 holder.txtBorrowFrom.setText("Người mượn:"+entity.getBorrowingUserName());
             }
-            holder.ratingBar.setNumStars(entity.getRateAvg());
+            holder.ratingBar.setNumStars((int)entity.getRateAvg());
             if (!Strings.isNullOrEmpty(entity.getImageId())) {
                 ClientUtils.setImage(holder.imgBook, R.drawable.ic_book_default, ClientUtils.getUrlImage(entity.getImageId(), Constance.IMAGE_SIZE_SMALL));
             }
